+++ beatmap/shared/environment.ts
@@ -31,12 +31,15 @@ export const EnvironmentRename: { [key in EnvironmentAllName]: string } = {
     TheSecondEnvironment: 'The Second',
     LizzoEnvironment: 'Lizzo',
     TheWeekndEnvironment: 'The Weeknd',
+    RockMixtapeEnvironment: 'Rock Mixtape',
     GlassDesertEnvironment: 'Glass Desert',
 } as const;
 
 /** List of available event type in environment. */
 export const EventList: {
-    [key in EnvironmentAllName]: Readonly<[ReadonlyArray<number>, ReadonlyArray<number>]>;
+    [key in EnvironmentAllName]: Readonly<
+        [ReadonlyArray<number>, ReadonlyArray<number>]
+    >;
 } = {
     DefaultEnvironment: [[0, 1, 2, 3, 4, 8, 9, 12, 13], []],
     OriginsEnvironment: [[0, 1, 2, 3, 4, 8, 9, 12, 13], []],
@@ -117,6 +120,49 @@ export const EventList: {
             37,
             38,
             40,
+        ],
+    ],
+    RockMixtapeEnvironment: [
+        [3, 4, 6],
+        [
+            0,
+            1,
+            2,
+            3,
+            4,
+            5,
+            6,
+            7,
+            8,
+            9,
+            10,
+            11,
+            12,
+            13,
+            14,
+            15,
+            16,
+            17,
+            18,
+            19,
+            20,
+            21,
+            22,
+            23,
+            24,
+            25,
+            26,
+            27,
+            28,
+            29,
+            30,
+            31,
+            32,
+            33,
+            34,
+            35,
+            36,
+            37,
         ],
     ],
 } as const;