+++ extensions/chroma/lightIDgenerate.ts
@@ -220,6 +220,14 @@ export const LightIDList: Record<
         3: Array.from(Array(7), (_, i) => i + 1),
         4: Array.from(Array(12), (_, i) => i + 1),
     },
+    // FIXME: unknown
+    RockMixtapeEnvironment: {
+        0: Array.from(Array(10), (_, i) => i + 1),
+        1: Array.from(Array(60), (_, i) => i + 1),
+        2: Array.from(Array(7), (_, i) => i + 1),
+        3: Array.from(Array(7), (_, i) => i + 1),
+        4: Array.from(Array(12), (_, i) => i + 1),
+    },
     GlassDesertEnvironment: {
         0: Array.from(Array(6), (_, i) => i + 1),
         1: Array.from(Array(8), (_, i) => i + 1),